1) виніс в окремий файл ренедер розмітки модалки 2) налаштував кнопки, які динамічно додають або видаляють зі сховища фільми 3) додав стилі цих кнопок
