+++ src/sass/layouts/_modal.scss
@@ -1,3 +1,12 @@
+.already-added {
+  background-color: #ff6b01;
+  color: #fff;
+
+  &:hover {
+    box-shadow: 0 8px 43px rgb(255 107 1 / 60%);
+  }
+}
+
 .modal {
   width: 806px;
   padding: 40px 12px;
